refactor scripts/finetune.py into new cli modules
